requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
packaging @ file:///home/task_176104874243446/conda-bld/packaging_1761049080023/work
python-dotenv==1.2.2
setuptools==80.10.2
sqlparse==0.5.5
wheel==0.46.3

